+++ litebird_sim/__init__.py
@@ -60,6 +60,15 @@ from .destriper import DestriperParameters, DestriperResult, destripe
 from .simulations import Simulation
 from .noise import add_noise
 from .scan_map import scan_map, scan_map_in_observations
+from .coordinates import DEFAULT_COORDINATE_SYSTEM, DEFAULT_TIME_SCALE
+from .spacecraft import (
+    compute_l2_pos_and_vel,
+    compute_lissajous_pos_and_vel,
+    spacecraft_pos_and_vel,
+    SpacecraftOrbit,
+    SpacecraftPositionAndVelocity,
+)
+from .dipole import add_dipole, add_dipole_to_observations, DipoleType
 from .version import __author__, __version__
 
 __all__ = [
@@ -140,4 +149,17 @@ __all__ = [
     #scan_map.py
     "scan_map",
     "scan_map_in_observations",
+    # dipole.py
+    "add_dipole",
+    "add_dipole_to_observations",
+    "DipoleType",
+    # coordinates.py
+    "DEFAULT_COORDINATE_SYSTEM",
+    "DEFAULT_TIME_SCALE",
+    # spacecraft.py
+    "compute_l2_pos_and_vel",
+    "compute_lissajous_pos_and_vel",
+    "spacecraft_pos_and_vel",
+    "SpacecraftOrbit",
+    "SpacecraftPositionAndVelocity",
 ]