+++ litebird_sim/observations.py
@@ -4,6 +4,7 @@ from typing import Union, List
 import astropy.time
 import numpy as np
 
+from .coordinates import DEFAULT_TIME_SCALE
 from .distribute import distribute_evenly
 
 # NOTE: When the deprecated pointing methods will be removed, the following
@@ -160,7 +161,7 @@ class Observation:
 
         if isinstance(self.start_time_global, astropy.time.Time):
             delta = astropy.time.TimeDelta(
-                1.0 / self.sampling_rate_hz, format="sec", scale="tdb"
+                1.0 / self.sampling_rate_hz, format="sec", scale=DEFAULT_TIME_SCALE
             )
         else:
             delta = 1.0 / self.sampling_rate_hz
@@ -516,6 +517,29 @@ class Observation:
         assert self.tod is None or len(info) == len(self.tod)
         setattr(self, name, info)
 
+    def get_delta_time(self) -> Union[float, astropy.time.TimeDelta]:
+        """Return the time interval between two consecutive samples in this observation
+
+        Depending whether the field ``start_time`` of the :class:`.Observation` object
+        is a ``float`` or a ``astropy.time.Time`` object, the return value is either a
+        ``float`` (in seconds) or an instance of ``astropy.time.TimeDelta``. See also
+        :meth:`.get_time_span`."""
+
+        delta = 1.0 / self.sampling_rate_hz
+        if isinstance(self.start_time, astropy.time.Time):
+            delta = astropy.time.TimeDelta(delta, format="sec", scale="tdb")
+
+        return delta
+
+    def get_time_span(self) -> Union[float, astropy.time.TimeDelta]:
+        """Return the temporal length of the current observation
+
+        This method can either return a ``float`` (in seconds) or a
+        ``astropy.time.TimeDelta`` object, depending whether the field ``start_time``
+        of the :class:`.Observation` object is a ``float`` or a
+        ``astropy.time.Time`` instance. See also :meth:`.get_delta_time`."""
+        return self.get_delta_time() * self.n_samples
+
     def get_times(self, normalize=False, astropy_times=False):
         """Return a vector containing the time of each sample in the observation
 
@@ -555,9 +579,7 @@ class Observation:
                 "to use astropy_times=True you must specify an astropy.time.Time "
                 "object in Observation.__init__"
             )
-            delta = astropy.time.TimeDelta(
-                1.0 / self.sampling_rate_hz, format="sec", scale="tdb"
-            )
+            delta = self.get_delta_time()
             return self.start_time + np.arange(self.n_samples) * delta
 
         if isinstance(self.start_time, astropy.time.Time):