+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # HEAD
 
+- Spacecraft simulator and dipole computation [#122](https://github.com/litebird/litebird_sim/pull/122)
+
 - Improve the way code is checked [#130](https://github.com/litebird/litebird_sim/pull/130)
 
 - Fix bugs [#126](https://github.com/litebird/litebird_sim/issues/126), [#124](https://github.com/litebird/litebird_sim/issues/124), [#120](https://github.com/litebird/litebird_sim/issues/120), [#111](https://github.com/litebird/litebird_sim/pull/111)